extract `Single` types to `tuple::single`
